Make DesktopComposeUiTest support IdlingResource registration
