+++ compose/ui/ui-test/api/desktop/ui-test.api
@@ -95,11 +95,9 @@ public abstract interface class androidx/compose/ui/test/ComposeUiTest : android
 	public abstract fun enableAccessibilityChecks ()V
 	public abstract fun getDensity ()Landroidx/compose/ui/unit/Density;
 	public abstract fun getMainClock ()Landroidx/compose/ui/test/MainTestClock;
-	public abstract fun registerIdlingResource (Landroidx/compose/ui/test/IdlingResource;)V
 	public abstract fun runOnIdle (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
 	public abstract fun runOnUiThread (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
 	public abstract fun setContent (Lkotlin/jvm/functions/Function2;)V
-	public abstract fun unregisterIdlingResource (Landroidx/compose/ui/test/IdlingResource;)V
 	public abstract fun waitForIdle ()V
 	public abstract fun waitUntil (Ljava/lang/String;JLkotlin/jvm/functions/Function0;)V
 	public static synthetic fun waitUntil$default (Landroidx/compose/ui/test/ComposeUiTest;Ljava/lang/String;JLkotlin/jvm/functions/Function0;ILjava/lang/Object;)V
@@ -129,6 +127,16 @@ public final class androidx/compose/ui/test/ComposeUiTest_skikoKt {
 	public static synthetic fun runInternalSkikoComposeUiTest$default (IILandroidx/compose/ui/unit/Density;Lkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/platform/PlatformContext$SemanticsOwnerListener;Lkotlinx/coroutines/test/TestDispatcher;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)V
 	public static final fun runSkikoComposeUiTest-Cqks5Fs (JLandroidx/compose/ui/unit/Density;Lkotlin/coroutines/CoroutineContext;Lkotlin/jvm/functions/Function1;)V
 	public static synthetic fun runSkikoComposeUiTest-Cqks5Fs$default (JLandroidx/compose/ui/unit/Density;Lkotlin/coroutines/CoroutineContext;Lkotlin/jvm/functions/Function1;ILjava/lang/Object;)V
+}
+
+public final class androidx/compose/ui/test/DesktopComposeUiTest : androidx/compose/ui/test/SkikoComposeUiTest {
+	public static final field $stable I
+	public fun <init> ()V
+	public fun <init> (IILkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/unit/Density;)V
+	public synthetic fun <init> (IILkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/unit/Density;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun registerIdlingResource (Landroidx/compose/ui/test/IdlingResource;)V
+	public final fun runDesktopTest (Lkotlin/jvm/functions/Function1;)V
+	public final fun unregisterIdlingResource (Landroidx/compose/ui/test/IdlingResource;)V
 }
 
 public abstract interface class androidx/compose/ui/test/DeviceConfigurationOverride {
@@ -638,13 +646,14 @@ public final class androidx/compose/ui/test/SemanticsSelector {
 	public final fun map (Ljava/lang/Iterable;Ljava/lang/String;)Landroidx/compose/ui/test/SelectionResult;
 }
 
-public final class androidx/compose/ui/test/SkikoComposeUiTest : androidx/compose/ui/test/ComposeUiTest {
+public class androidx/compose/ui/test/SkikoComposeUiTest : androidx/compose/ui/test/ComposeUiTest {
 	public static final field $stable I
 	public field scene Landroidx/compose/ui/scene/ComposeScene;
 	public fun <init> (IILkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/unit/Density;)V
 	public synthetic fun <init> (IILkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/unit/Density;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public fun <init> (IILkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/unit/Density;Landroidx/compose/ui/platform/PlatformContext$SemanticsOwnerListener;Lkotlinx/coroutines/test/TestDispatcher;)V
 	public synthetic fun <init> (IILkotlin/coroutines/CoroutineContext;Landroidx/compose/ui/unit/Density;Landroidx/compose/ui/platform/PlatformContext$SemanticsOwnerListener;Lkotlinx/coroutines/test/TestDispatcher;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	protected fun areAllResourcesIdle ()Z
 	public fun awaitIdle (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public final fun captureToImage ()Landroidx/compose/ui/graphics/ImageBitmap;
 	public final fun captureToImage (Landroidx/compose/ui/semantics/SemanticsNode;)Landroidx/compose/ui/graphics/ImageBitmap;
@@ -656,13 +665,11 @@ public final class androidx/compose/ui/test/SkikoComposeUiTest : androidx/compos
 	public final fun getScene ()Landroidx/compose/ui/scene/ComposeScene;
 	public fun onAllNodes (Landroidx/compose/ui/test/SemanticsMatcher;Z)Landroidx/compose/ui/test/SemanticsNodeInteractionCollection;
 	public fun onNode (Landroidx/compose/ui/test/SemanticsMatcher;Z)Landroidx/compose/ui/test/SemanticsNodeInteraction;
-	public fun registerIdlingResource (Landroidx/compose/ui/test/IdlingResource;)V
 	public fun runOnIdle (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
 	public fun runOnUiThread (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
 	public final fun runTest (Lkotlin/jvm/functions/Function1;)Ljava/lang/Object;
 	public fun setContent (Lkotlin/jvm/functions/Function2;)V
 	public final fun setScene (Landroidx/compose/ui/scene/ComposeScene;)V
-	public fun unregisterIdlingResource (Landroidx/compose/ui/test/IdlingResource;)V
 	public fun waitForIdle ()V
 	public fun waitUntil (Ljava/lang/String;JLkotlin/jvm/functions/Function0;)V
 }
