+++ compose/ui/ui-test/src/desktopMain/kotlin/androidx/compose/ui/test/ComposeUiTest.desktop.kt
@@ -16,11 +16,9 @@
 
 package androidx.compose.ui.test
 
+import androidx.compose.ui.unit.Density
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
-
-@OptIn(ExperimentalTestApi::class)
-typealias DesktopComposeUiTest = SkikoComposeUiTest
 
 /**
  * Variant of [runComposeUiTest] that allows you to specify the size of the surface.
@@ -38,5 +36,40 @@ fun runDesktopComposeUiTest(
     effectContext: CoroutineContext = EmptyCoroutineContext,
     block: DesktopComposeUiTest.() -> Unit
 ) {
-    DesktopComposeUiTest(width, height, effectContext).runTest(block)
+    DesktopComposeUiTest(width, height, effectContext).runDesktopTest(block)
+}
+
+@ExperimentalTestApi
+class DesktopComposeUiTest internal constructor(
+    width: Int = 1024,
+    height: Int = 768,
+    effectContext: CoroutineContext = EmptyCoroutineContext,
+    density: Density = Density(1f),
+) : SkikoComposeUiTest(width, height, effectContext, density) {
+
+    private val idlingResources = mutableSetOf<IdlingResource>()
+
+    override fun areAllResourcesIdle(): Boolean {
+        return synchronized(idlingResources) {
+            idlingResources.all { it.isIdleNow }
+        }
+    }
+
+    fun registerIdlingResource(idlingResource: IdlingResource) {
+        synchronized(idlingResources) {
+            idlingResources.add(idlingResource)
+        }
+    }
+
+    fun unregisterIdlingResource(idlingResource: IdlingResource) {
+        synchronized(idlingResources) {
+            idlingResources.remove(idlingResource)
+        }
+    }
+
+    fun runDesktopTest(block: DesktopComposeUiTest.() -> Unit) {
+        runTest {
+            block()
+        }
+    }
 }
